+++ src/components/CheckboxList/checkbox-list-props.ts
@@ -1,0 +1,11 @@
+export interface CheckboxOption {
+  id: string;
+  value: string;
+  checked: boolean;
+}
+
+export interface CheckboxListProps {
+  options: CheckboxOption[];
+  disabled: boolean; // If disable all checkboxes
+  onChange: (data: CheckboxOption) => void;
+}