+++ src/components/Switch/switch-props.tsx
@@ -1,0 +1,6 @@
+import { Assign, BoxOwnProps } from 'theme-ui';
+
+export interface SwitchProps
+  extends Assign<React.ComponentProps<'input'>, BoxOwnProps> {
+  label?: string;
+}